% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities_Cluster_image.R
\name{Cleavage_rules_fun}
\alias{Cleavage_rules_fun}
\title{Cleavage_rules_fun}
\usage{
Cleavage_rules_fun()
}
\value{
a table of Cleavage rules
}
\description{
This is a function will return a list of pre-defined enzyme digestion specificity.
}
\examples{
Cleavage_rules_fun()

}
