% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities_Cluster_image.R
\name{cluster_image_grid}
\alias{cluster_image_grid}
\title{cluster_image_grid}
\usage{
cluster_image_grid(
  clusterID,
  SMPLIST,
  imdata,
  ClusterID_colname = "Protein",
  componentID_colname = "Peptide",
  Component_plot_threshold = 2,
  Component_plot_coloure = c("mono", "as.cluster"),
  smooth.image = "gaussian",
  contrast.enhance = "suppression",
  colorpallet = "Set1",
  plot_layout = c("line", "grid"),
  export_Header_table = F,
  export_footer_table = F,
  plot_style = c("fleximaging", "ClusterOnly", "rainbow"),
  protein_coverage = F,
  footer_style = "Length",
  output_png_width_limit = 1980,
  attach_summary_cluster = T,
  cluster_color_scale = c("blackwhite", "fleximaging"),
  remove_cluster_from_grid = T,
  img_brightness = 100,
  ppm = 20,
  list_of_protein_sequence,
  workdir = getwd(),
  pixel_size_um = 50,
  Score_thres = NULL
)
}
\description{
This function renders the clustered images for maldi imaging data set
}
