% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Canidate_processing.R
\name{get_atoms}
\alias{get_atoms}
\title{get_atoms}
\usage{
get_atoms(Symbol)
}
\arguments{
\item{Symbol}{the Symbol of the molecule}
}
\value{
a atoms list
}
\description{
This is a function that prepare the atoms list for candidate molecules.
}
\examples{
get_atoms(Symbol="C7H13O8P")

}
