% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities_Cluster_image.R
\name{grid.ftable}
\alias{grid.ftable}
\title{grid.ftable}
\usage{
grid.ftable(d, padding = unit(4, "mm"), ...)
}
\value{
a table image
}
\description{
This is a function will plot table.
}
\examples{
grid.ftable()

}
