% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{package}
\name{HiTMaP}
\alias{HiTMaP}
\alias{Generatespectrum_workflow_maldiquant}
\title{\tabular{ll}{
Package: \tab HiTMaP \cr
License: \tab GPL (>= 2)\cr
URL: \tab 
}}
\usage{
Generatespectrum_workflow_maldiquant()
}
\description{
\tabular{ll}{
Package: \tab HiTMaP \cr
License: \tab GPL (>= 2)\cr
URL: \tab 
}
}
\references{
TBD
}
\author{
George GUO \email{George.GUO@auckland.ac.nz}
}
\keyword{package}
