% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyAPP_GUI.R
\docType{function}
\name{HiTMaP_GUI}
\alias{HiTMaP_GUI}
\title{HiTMaP shiny-based graphical user interface (GUI)}
\usage{
HiTMaP_GUI(wd = "~/", port = 3838)
}
\description{
HiTMaP shiny-based graphical user interface (GUI)
}
\examples{
## Start the GUI
\dontrun{
HiTMaP_GUI()
}



}
