% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peptidedigest.R
\name{imaging_Spatial_Quant}
\alias{imaging_Spatial_Quant}
\title{imaging_Spatial_Quant}
\usage{
imaging_Spatial_Quant(datafile = tk_choose.files(filter =
  matrix(c("imzml file", ".imzML", "Text", ".txt", "All files", "*"), 3, 2,
  byrow = TRUE), caption =
  "Choose single or multiple file(s) for analysis"), threshold = 0,
  ppm = 2.5, Quant_list = "Metabolites of Interest.csv",
  adducts = c("M-H", "M+Cl"), cal.mz = F, mzlist_bypass = T,
  PMF_analysis = TRUE, Protein_feature_summary = T,
  plot_cluster_image = T, plot_style = "fleximaging",
  Peptide_feature_summary = T, plot_ion_image = FALSE,
  parallel = detectCores()/2, spectra_segments_per_file = 5,
  spatialKMeans = F, Smooth_range = 1, Virtual_segmentation = T,
  Virtual_segmentation_rankfile = tk_choose.files(default =
  "Z:/George skyline results/maldiimaging/Maldi_imaging - Copy/radius_rank.csv",
  caption = "Choose Virtual segmentation rank info file"),
  Spectrum_feature_summary = T, Region_feature_summary = T,
  Region_feature_analysis = T, plot_each_metabolites = T,
  Cluster_level = "High", ClusterID_colname = "Name",
  Region_feature_analysis_bar_plot = T, norm_datafiles = T,
  norm_Type = "Median", mzrange = NULL, BPPARAM = bpparam(),
  rotateimg = NULL, ...)
}
\arguments{
\item{threshold}{specify the intensities threshold (0 to 1 in percentage)to report a identified molecule}

\item{ppm}{the mz tolerance (in ppm) for peak integration}

\item{Quant_list}{the quantifiaction candidate list, spatial quantification will go through every datafile and collect the ion intensities for each listed component}

\item{adducts}{the adducts list to be used for generating the PMF search candidates}

\item{cal.mz}{If set with \code{"true"}, the function will recalculate the mz value according to the column named "formular" in the \code{Quant_list} and the specified adducts.}

\item{mzlist_bypass}{Set \code{"true"} if you want to bypass the mzlist generating process}

\item{Protein_feature_summary}{\code{"PMF_analysis"} follow-up process that will collect all the identified peptide information and associate them with possible proteins}

\item{plot_cluster_image}{\code{"Protein_feature_summary"} follow-up process that will plot the protein cluster image}

\item{plot_ion_image}{\code{"Peptide_feature_summarya"} follow-up process that will plot every connponents in the \code{"peptide shortlist"}}

\item{parallel}{the number of threads will be used in the PMF search, this option now only works for windows OS}

\item{spectra_segments_per_file}{optimal number of distinctive regions in the imaging, a virtual segmentation will be applied to the image files with this value. To have a better PMF result you may set a value that in the sweet point of sensitivety and false discovery rate (FDR).}

\item{spatialKMeans}{set true to enable a \code{"spatialKMeans"}  method for the automatic virtual segmentation. If a region rank file was supplied, you can disable this to perform a mannual segmentation.}

\item{Smooth_range}{\code{"spatialKMeans"} pixel smooth range}

\item{Virtual_segmentation}{set \code{"TRUE"} if you want to overide the automaitic segmentation}

\item{Virtual_segmentation_rankfile}{specify a region rank file contains region information for manualy region segmentation}

\item{Peptide_feature_summarya}{\code{"PMF_analysis"} follow-up process that will summarize all datafiles identified peptides and generats a \code{"peptide shortlist"} in the result summary folder}
}
\value{
None
}
\description{
This is a spatial quantitation function for maldi imaging data set
this function will read the candidate list file and generate quantification result
}
\examples{
imaging_Spatial_Quant(threshold=0.05, ppm=5,Digestion_site="[G]",
                       missedCleavages=0:1,Fastadatabase="murine_matrisome.fasta",
                       adducts=c("M+H","M+NH4","M+Na"),PMF_analysis=TRUE,
                       Protein_feature_summary=TRUE,plot_cluster_image=TRUE,
                       Peptide_feature_summary=TRUE,plot_ion_image=FALSE,
                       parallel=3,spectra_segments_per_file=5,spatialKMeans=TRUE
                       )

}
