% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Canidate_processing.R
\name{merge_atoms}
\alias{merge_atoms}
\title{merge_atoms}
\usage{
merge_atoms(
  atoms,
  addelements,
  check_merge = F,
  mode = c("add", "ded"),
  multiplier = c(1, 1)
)
}
\arguments{
\item{atoms}{the file name of candidate list}

\item{addelements}{the folder that contains candidate list}

\item{check_merge}{Check the final formula for any negative number of atom(s).}

\item{mode}{add the elements to the formula by using "add", and deduct the elements by using "ded".}

\item{multiplier}{Two number to define the atom and adding elements' proportion respectively.}
}
\value{
an atom list
}
\description{
This is a function that merge the atoms to a melecule atom list .
}
\examples{
merge_atoms(atoms=get_atoms("C7H13O8P"),addelements=list(H=1),check_merge=T,mode="ded",multiplier=c(1,2))

}
