% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Canidate_processing.R
\name{Meta_feature_list_fun}
\alias{Meta_feature_list_fun}
\title{Meta_feature_list_fun}
\usage{
Meta_feature_list_fun(database, workdir = getwd(), adducts = adducts,
  cal.mz = TRUE, bypass = FALSE, BPPARAM = bpparam())
}
\arguments{
\item{database}{the file name of candidate list}

\item{workdir}{the folder that contains candidate list}

\item{adducts}{the adducts list to be used for generating the PMF search candidates}

\item{cal.mz}{If set with \code{"TRUE"}, the function will recalculate the mz value according to the column named "formula" in the \code{database} and the specified adducts.}

\item{BPPARAM}{parallel processing parameter for BiocParallel}

\item{mzlist_bypass}{Set \code{"TRUE"} if you want to bypass the mzlist generating process, the function will keep the mz and adduct as it is for the furture analysis. Be sure that the table contains "mz", "adduct" and "moleculeNames" as they are essential for later steps.}
}
\value{
a table of candiate list
}
\description{
This is a function that prepare the candiate list for maldi imaging data qualitative or quantitative analysis.
this function will read the candidate list file and generate mz for the adducts list defined in \code{"adducts"}.
}
\examples{
Meta_feature_list_fun(database="lipid candidates.csv",adducts=c("M-H","M+Cl"))

}
