% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Canidate_processing.R
\name{Peptide_modification}
\alias{Peptide_modification}
\title{Peptide_modification}
\usage{
Peptide_modification(retrive_ID = NULL, mod_position = NULL, update_unimod = F)
}
\arguments{
\item{retrive_ID}{the file name of candidate list}

\item{mod_position}{the folder that contains candidate list}

\item{update_unimod}{the adducts list to be used for generating the PMF search candidates}
}
\value{
a table of modification list
}
\description{
This is a function that prepare the modification list for maldi imaging data qualitative or quantitative analysis.
this function will load a pre-built Unimod database (unimod.df) into global environment.
}
\examples{
Peptide_modification()

}
