% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Canidate_processing.R
\name{Protein_feature_list_fun}
\alias{Protein_feature_list_fun}
\title{Protein_feature_list_fun}
\usage{
Protein_feature_list_fun(workdir = getwd(), database,
  Digestion_site = "[G]", missedCleavages = 0:1, adducts = c("M+H",
  "M+NH4", "M+Na"), BPPARAM = bpparam(), Decoy_adducts = c("M+He",
  "M+Ne", "M+Ar", "M+Kr", "M+Xe", "M+Rn"), Decoy_mode = c("adducts",
  "elements", "isotope"), Decoy_search = T, mzrange = c(700, 4000),
  output_candidatelist = T, use_previous_candidates = F)
}
\arguments{
\item{workdir}{the folder that contains fasta file}

\item{database}{the file name of fasta file}

\item{Digestion_site}{Digestion pattern, this is the}

\item{missedCleavages}{Define a number range for misscleaved peptides to be considered}

\item{adducts}{the adducts list to be used for generating the PMF search candidates}

\item{BPPARAM}{parallel processing parameter for BiocParallel}
}
\value{
a table of peptide candiate list
}
\description{
This is a function that prepare the candiate list for maldi imaging data qualitative or quantitative analysis.
this function will read the fasta file and generate mz for the adducts list defined in \code{"adducts"}.
}
\examples{
Protein_feature_list_fun(database="lipid candidates.csv",adducts=c("M+H","M+NH4","M+Na"))

}
