% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workflow.R
\name{imaging_identification}
\alias{imaging_identification}
\title{imaging_identification}
\usage{
imaging_identification(
  datafile,
  projectfolder = NULL,
  threshold = 0.001,
  ppm = 5,
  mode = c("Proteomics", "Metabolomics"),
  Digestion_site = "trypsin",
  missedCleavages = 0:1,
  Fastadatabase = "uniprot-bovin.fasta",
  adducts = c("M+H"),
  Modifications = list(fixed = NULL, fixmod_position = NULL, variable = NULL,
    varmod_position = NULL),
  Substitute_AA = NULL,
  Decoy_search = TRUE,
  Decoy_adducts = c("M+ACN+H", "M+IsoProp+H", "M+DMSO+H", "M+Co", "M+Ag", "M+Cu",
    "M+He", "M+Ne", "M+Ar", "M+Kr", "M+Xe", "M+Rn"),
  Decoy_mode = "isotope",
  mzrange = c(700, 4000),
  Database_stats = F,
  adjust_score = FALSE,
  IMS_analysis = TRUE,
  PMFsearch = IMS_analysis,
  Load_candidatelist = IMS_analysis || plot_cluster_image_grid,
  Bypass_generate_spectrum = FALSE,
  peptide_ID_filter = 2,
  Protein_feature_summary = TRUE,
  Peptide_feature_summary = TRUE,
  plot_ion_image = FALSE,
  parallel = detectCores(),
  spectra_segments_per_file = 4,
  Segmentation = c("spatialKMeans", "spatialShrunkenCentroids", "Virtual_segmentation",
    "none", "def_file"),
  Segmentation_def = "Segmentation_def.csv",
  Segmentation_ncomp = "auto-detect",
  Segmentation_variance_coverage = 0.8,
  preprocess = list(force_preprocess = FALSE, use_preprocessRDS = TRUE, smoothSignal =
    list(method = "disable"), reduceBaseline = list(method = "locmin"), peakPick =
    list(method = "adaptive"), peakAlign = list(tolerance = ppm/2, units = "ppm"),
    normalize = list(method = c("rms", "tic", "reference")[1], mz = 1)),
  Smooth_range = 1,
  Virtual_segmentation_rankfile = NULL,
  Rotate_IMG = NULL,
  Region_feature_summary = FALSE,
  Spectrum_validate = TRUE,
  output_candidatelist = TRUE,
  use_previous_candidates = FALSE,
  score_method = "SQRTP",
  plot_cluster_image_grid = FALSE,
  plot_cluster_image_maxretry = 2,
  plot_cluster_image_overwrite = F,
  smooth.image = "gaussian",
  componentID_colname = "Peptide",
  ClusterID_colname = "Protein",
  Protein_desc_of_interest = ".",
  Protein_desc_of_exclusion = NULL,
  plot_unique_component = TRUE,
  FDR_cutoff = 0.05,
  use_top_rank = NULL,
  plot_matching_score = F,
  Component_plot_coloure = "mono",
  cluster_color_scale = "blackwhite",
  plot_layout = "line",
  export_Header_table = T,
  export_footer_table = T,
  attach_summary_cluster = T,
  remove_cluster_from_grid = attach_summary_cluster,
  pixel_size_um = 50,
  img_brightness = 100,
  Thread = 4,
  cluster_rds_path = NULL,
  remove_score_outlier = F,
  Plot_score_IQR_cutoff = 0.75,
  Plot_score_abs_cutoff = -0.1,
  ...
)
}
\arguments{
\item{datafile}{the data files' path for the analysis, leave it as blank to enable a graphical user interface to select the data}

\item{projectfolder}{optional, if NULL script will extract the path from datafile(s), and use the first workdir as project folder}

\item{threshold}{specify the intensities threshold (0 to 1 in percentage)to report a identified molecule}

\item{ppm}{the mz tolerance (in ppm) for peak integration}

\item{Digestion_site}{Set the enzyme digestion specificity by one or more regex expressions or the name of a enzyme}

\item{missedCleavages}{miss cleavage number allowed in this PMF search}

\item{Fastadatabase}{the fasta database used in this pmf search, the file should be placed in the same folder with data files}

\item{adducts}{the adducts list to be used for generating the PMF search candidates}

\item{Modifications}{set the modifications}

\item{Substitute_AA}{set the amino acid Substitutions}

\item{Decoy_search}{enable (default) or disable the decoy search}

\item{Decoy_adducts}{define the adduct list for decoy search. the decoy adducts could be "M+ACN+H","M+IsoProp+H","M+DMSO+H","M+Co","M+Ag","M+Cu","M+He","M+Ne","M+Ar","M+Kr","M+Xe" or"M+Rn".}

\item{Decoy_mode}{select the decoy search mode between "isotope" (default), "element" and "adduct"}

\item{mzrange}{define the mz range for the experiment, default is 700 to 4000 m/z.}

\item{IMS_analysis}{Set \code{"true"} if you want to perform data pre-processing and proteomics search, set \code{"false"} if you want to bypass it}

\item{peptide_ID_filter}{set the minimal count of peptides needed to identify a protein}

\item{Protein_feature_summary}{\code{"IMS_analysis"} follow-up process that will collect all the identified peptide information and associate them with possible proteins}

\item{Peptide_feature_summary}{\code{"IMS_analysis"} follow-up process that will summarize all datafiles identified peptides and generats a \code{"peptide shortlist"} in the result summary folder}

\item{plot_ion_image}{\code{"Peptide_feature_summarya"} follow-up process that will plot every connponents in the \code{"peptide shortlist"}. please use the cluster image grid to output the images.}

\item{parallel}{the number of threads will be used in the PMF search, this option now only works for windows OS}

\item{spectra_segments_per_file}{optimal number of distinctive regions in the tissue section, a virtual segmentation will be applied to the image files with this value. To have a better PMF result you may set a value that in the sweet point of sensitivety and false discovery rate (FDR).}

\item{Segmentation}{set as "spatialKMeans" to enable a \code{"spatialKMeans"} Segmentation; set as "spatialShrunkenCentroids" to enable a \code{"spatialShrunkenCentroids"} Segmentation; If a region rank file was supplied, you can set this as "Virtual_segmentation" to perform a manual segmentation; Set it as "none" to bypass the segmentation.}

\item{preprocess}{a list of params that define the IMS data pre-processing procedure}

\item{Smooth_range}{\code{"Segmentation"} pixel smooth range}

\item{Virtual_segmentation_rankfile}{specify a region rank file contains region information for manualy region segmentation}

\item{Rotate_IMG}{specify a configuration file to further change the rotation of the images}

\item{Region_feature_summary}{\code{"IMS_analysis"} follow-up process that will summarize mz feature of all regions of all data files into the summary folder}

\item{use_previous_candidates}{set as TRUE to reload the previously generated candidate list.}

\item{score_method}{specify the peptide spectrum scoring method, "SQRTP" is recommended.}

\item{plot_cluster_image_grid}{set as \code{"TRUE"} to enable the protein cluster image function.}

\item{plot_cluster_image_overwrite}{Set as true to generate the cluster images regardless the existance of previously file(s)}

\item{componentID_colname}{Specify the component ID column in the result spreadsheet.}

\item{ClusterID_colname}{Specify the cluster ID column in the result spreadsheet.}

\item{Protein_desc_of_interest}{Specify a list of protein descriptions for cluster image plotting. Default setting will plot all reported proteins.}

\item{Protein_desc_of_exclusion}{Specify a list of protein descriptions to be excluded from cluster image plotting.}

\item{plot_unique_component}{Set as \code{"TRUE"} to plot only the unique components in the cluster image plotting.}

\item{FDR_cutoff}{set the protein FDR cutoff threshold, default is 5 percent}

\item{plot_matching_score}{enable the spectrum matching overlay plot}

\item{Component_plot_coloure}{set as "mono" to use a pre-defined color scale to plot component images. Set as "as.cluster" to use the previously assigned mono color in the additive cluster binning process.}

\item{cluster_color_scale}{Set as "blackwhite" to use only black and white color in the cluster image plotting. using "blackwhite" in cluster_color_scale will overwrite the components' color setting.}

\item{plot_layout}{Set as \code{"line"} to plot cluster and component images for multiple data file or as \code{"grid"} to plot cluster images for single data file. In "grid" mode, Image's will be rendered into a grid with 5 columns.}

\item{export_Header_table}{Set as \code{"TRUE"} to plot the header in the cluster image plotting. Header table includes the basic information of cluster and components.}

\item{export_footer_table}{Set as \code{"TRUE"} to plot the footer in the cluster image plotting. Footer shows the protein coverage in the Proteomics mode.}

\item{attach_summary_cluster}{Set as \code{"TRUE"} to attach an enlarged cluster image to the bottom of the cluster image.}

\item{remove_cluster_from_grid}{Set as \code{"TRUE"} to remove the cluster image from the cluster image grid. it is recommended to set this same as the attach_summary_cluster.}

\item{cluster_rds_path}{set as NULL if there is not preprocessed.rds available for a single file, script will load the raw data file which may reduce the signal intensities. For multiple samples, scripts will try to load the RDS file from each "ID" folder and merge the mz features via instrument resolution setting and output a combined RDS file to the project folder. For multiple files cluster images rendering user should set the attach_summary_cluster as False, and set remove_cluster_from_grid as true.}
}
\value{
None
}
\description{
This is a peptide mass fingerprint search function for maldi imaging data analysis
}
\examples{
imaging_identification(threshold=0.05, ppm=5,Digestion_site="[G]",
                       missedCleavages=0:1,Fastadatabase="murine_matrisome.fasta",
                       adducts=c("M+H","M+NH4","M+Na"),IMS_analysis=TRUE,
                       Protein_feature_summary=TRUE,plot_cluster_image=TRUE,
                       Peptide_feature_summary=TRUE,plot_ion_image=FALSE,
                       parallel=3,spectra_segments_per_file=5,Segmentation="spatialKMeans"
                       )

}
