% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peptidedigest.R
\name{imaging_identification}
\alias{imaging_identification}
\title{imaging_identification}
\usage{
imaging_identification(datafile, threshold = 0.005, ppm = 5,
  mode = c("Proteomics", "Metabolomics"), Digestion_site = "[GN][LI]",
  missedCleavages = 0:1, Fastadatabase = "murine_matrisome.fasta",
  adducts = c("M+H"), Decoy_search = T, Decoy_adducts = c("M+ACN+H",
  "M+IsoProp+H", "M+DMSO+H", "M+Co", "M+Ag", "M+Cu", "M+He", "M+Ne",
  "M+Ar", "M+Kr", "M+Xe", "M+Rn"), Decoy_mode = "isotope",
  adjust_score = F, PMF_analysis = TRUE,
  Bypass_generate_spectrum = F, Protein_feature_summary = TRUE,
  plot_cluster_image = TRUE, ClusterID_colname = "Protein",
  Peptide_feature_summary = TRUE, plot_ion_image = FALSE,
  parallel = detectCores(), spectra_segments_per_file = 5,
  spatialKMeans = TRUE, Smooth_range = 1,
  Virtual_segmentation = FALSE, Virtual_segmentation_rankfile = NULL,
  rotateimg = NULL, Region_feature_summary = F,
  Spectrum_validate = T, output_candidatelist = T,
  use_previous_candidates = F, score_method = "SQRTP",
  plot_cluster_image_grid = F, ...)
}
\arguments{
\item{threshold}{specify the intensities threshold (0 to 1 in percentage)to report a identified molecule}

\item{ppm}{the mz tolerance (in ppm) for peak integration}

\item{Digestion_site}{the enzyme digestion site specificity}

\item{missedCleavages}{miss cleavage number allowed in this PMF search}

\item{Fastadatabase}{the fasta database used in this pmf search, the file should be placed in the same folder with data files}

\item{adducts}{the adducts list to be used for generating the PMF search candidates}

\item{PMF_analysis}{Set \code{"true"} if you want to have a PMF search, set \code{"false"} if you want to bypass it}

\item{Protein_feature_summary}{\code{"PMF_analysis"} follow-up process that will collect all the identified peptide information and associate them with possible proteins}

\item{plot_cluster_image}{\code{"Protein_feature_summary"} follow-up process that will plot the protein cluster image}

\item{plot_ion_image}{\code{"Peptide_feature_summarya"} follow-up process that will plot every connponents in the \code{"peptide shortlist"}}

\item{parallel}{the number of threads will be used in the PMF search, this option now only works for windows OS}

\item{spectra_segments_per_file}{optimal number of distinctive regions in the imaging, a virtual segmentation will be applied to the image files with this value. To have a better PMF result you may set a value that in the sweet point of sensitivety and false discovery rate (FDR).}

\item{spatialKMeans}{set true to enable a \code{"spatialKMeans"}  method for the automatic virtual segmentation. If a region rank file was supplied, you can disable this to perform a mannual segmentation.}

\item{Smooth_range}{\code{"spatialKMeans"} pixel smooth range}

\item{Virtual_segmentation}{set \code{"TRUE"} if you want to overide the automaitic segmentation}

\item{Virtual_segmentation_rankfile}{specify a region rank file contains region information for manualy region segmentation}

\item{Peptide_feature_summarya}{\code{"PMF_analysis"} follow-up process that will summarize all datafiles identified peptides and generats a \code{"peptide shortlist"} in the result summary folder}
}
\value{
None
}
\description{
This is a peptide mass fingerprint search function for maldi imaging data analysis
}
\examples{
imaging_identification(threshold=0.05, ppm=5,Digestion_site="[G]",
                       missedCleavages=0:1,Fastadatabase="murine_matrisome.fasta",
                       adducts=c("M+H","M+NH4","M+Na"),PMF_analysis=TRUE,
                       Protein_feature_summary=TRUE,plot_cluster_image=TRUE,
                       Peptide_feature_summary=TRUE,plot_ion_image=FALSE,
                       parallel=3,spectra_segments_per_file=5,spatialKMeans=TRUE
                       )

}
